requirements.txt
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
peewee==3.15.4
python-telegram-bot-pagination==0.0.3
requests==2.32.2
